+++ backend/src/main/java/ru/mephi/trainer/rest/dto/response/TrainerProgressResponse.java
@@ -15,8 +15,8 @@ import java.util.UUID;
 public class TrainerProgressResponse {
     private UUID trainerId;
     private String trainerName;
-    private Integer earnedScore;      // заработал
-    private Integer maxPossibleScore; // максимум можно
+    private Double earnedScore;      // заработал
+    private Double maxPossibleScore; // максимум можно
     private Integer tasksCompleted;   // решено задач
     private Integer totalTasks;       // всего задач в тренажёре
     private List<CompletedTaskTrainerPointResponse> tasksInTrainer;  // баллы за тренажер
